+++ Server/routes/meals.rb
@@ -1,6 +1,14 @@
 # http://localhost:4567/meals/1
 get '/meals/:id' do
-  verifyLoggedIn
+  message = verifyLoggedIn
+  if message
+    status 400
+    res = { message: message }
+    content_type :json
+    return res.to_json
+  end
+  
+  
   results_array = Array.new
   
   query = "SELECT * FROM Meal WHERE id=#{params[:id]}"
@@ -16,7 +24,15 @@ end
 
 # http://localhost:4567/meals
 post '/meals' do
-  verifyLoggedIn
+  message = verifyLoggedIn
+  if message
+    status 400
+    res = { message: message }
+    content_type :json
+    return res.to_json
+  end
+  
+  
   unless request.body.read.empty?
     request.body.rewind # because "read" method is IO thing and the unless condition is ending it
     data = JSON.parse request.body.read
@@ -66,7 +82,15 @@ end
 
 # http://localhost:4567/meals/1
 put '/meals/:id' do
-  verifyLoggedIn
+  message = verifyLoggedIn
+  if message
+    status 400
+    res = { message: message }
+    content_type :json
+    return res.to_json
+  end
+  
+  
   unless request.body.read.empty?
     request.body.rewind # because "read" method is IO thing and the unless condition is ending it
     data = JSON.parse request.body.read
@@ -103,7 +127,15 @@ end
 
 # http://localhost:4567/meals/1
 delete '/meals/:id' do
-  verifyLoggedIn
+  message = verifyLoggedIn
+  if message
+    status 400
+    res = { message: message }
+    content_type :json
+    return res.to_json
+  end
+  
+  
   query = "DELETE FROM Item_has_Meal WHERE Meal_id=#{params[:id]}"
   @@mysqlclient.query(query, as: :hash)
